Change setMovePower parameter type to float
Updated the setMovePower method to accept a float instead of an int for motor power, allowing for more precise control of motor speed.

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/DriveTrain.java
@@ -204,7 +204,7 @@ public class DriveTrain {
         backRightMotor.setPower(-power);
     }
 
-    private void setMovePower(int power) {
+    private void setMovePower(float power) {
         // Set the power for all motors
         frontLeftMotor.setPower(power);
         frontRightMotor.setPower(power);
